amore_crawler/requirements.txt
python-dotenv>=1.0.0
pyyaml>=6.0.1
requests>=2.31.0
tenacity>=8.2.3
pandas>=2.1.0
pyarrow>=14.0.0
tqdm>=4.66.0

selenium>=4.20.0
webdriver-manager>=4.0.2

beautifulsoup4>=4.12.2
lxml>=5.1.0
